requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2024.12.14
defusedxml==0.7.1
distro==1.9.0
fastjsonschema==2.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.2
jiter==0.8.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
MarkupSafe==2.1.3
mistune==3.0.1
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
openai==1.59.3
packaging==23.1
pandocfilters==1.5.0
platformdirs==3.10.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pyzmq==25.1.1
referencing==0.30.2
rpds-py==0.10.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.67.1
traitlets==5.9.0
typing_extensions==4.12.2
webencodings==0.5.1
